app/requirements.txt
pure-python-adb
opencv-python
pillow
pytesseract
python-dotenv
openai>=1.0.0
google-generativeai
flask
facebook-wda
pymobiledevice3
boto3